+++ geompp_python/src/bind_polygon3d.cpp
@@ -23,6 +23,7 @@ void bind_polygon3d(py::module_& m) {
         .def("area",        &geompp::Polygon3D::Area)
         .def("perimeter",   &geompp::Polygon3D::Perimeter)
         .def("is_simple",   &geompp::Polygon3D::IsSimple,    "Returns True if the polygon boundary has no self-intersections.")
+        .def("is_convex",   &geompp::Polygon3D::IsConvex)
         .def("distance_to", &geompp::Polygon3D::DistanceTo,  "point"_a)
         .def("contains",       &geompp::Polygon3D::Contains,      "point"_a)
         .def("is_on_boundary", &geompp::Polygon3D::IsOnBoundary, "point"_a)